Pax-main/requirements.txt
fastapi
uvicorn
discord.py
asyncpg
deep-translator
aiohttp
